The default command wasn't recognized, extended case matching
There is (very likely) a better solution to this problem. The problem
originates in orml.bash where 'shift 2' is used, the comment there
explains why we shift twice. In a scenario where we shift only once
there, the default command would be recognized. I probably won't ever return
for this problem but it is worth mentioning.